Namespace Hooks module to avoid naming collision
Our own top-level Hooks module collides with that provided in the
hooks gem, now an indirect dependency (needed by livingstyleguide).
