+++ lib/open_project/hooks.rb
@@ -26,6 +26,11 @@
 # See doc/COPYRIGHT.rdoc for more details.
 #++
 
-class Hooks::ActivityIndexHeadHook < Redmine::Hook::ViewListener
-  render_on :activity_index_head, :partial => 'hooks/timelines/activity_index_head'
+##
+# A placeholder module which contains OpenProject hooks
+module OpenProject
+  module Hooks; end
 end
+
+# actual hooks are added with the following require statemens
+require 'open_project/hooks/view_account_login_auth_provider'
