+++ lib/open_project/hooks/view_projects_form_hook.rb
@@ -26,11 +26,10 @@
 # See doc/COPYRIGHT.rdoc for more details.
 #++
 
-module Hooks
-  ##
-  # Hook called in the login forms which displays the different auth providers
-  class ViewAccountLoginAuthProviderHook < Redmine::Hook::ViewListener
-    render_on :view_account_login_auth_provider,
-              :partial => 'hooks/login/auth_provider'
+module OpenProject
+  class Hooks::ViewProjectsFormHook < Redmine::Hook::ViewListener
+    render_on :view_projects_form,
+              :partial => 'hooks/timelines/view_projects_form'
+  
   end
 end
